\name{MKclass-package}
\alias{MKclass-package}
\alias{MKclass}
\docType{package}
\title{
Functions for Statistical Classification.
}
\description{
Contains several functions for statistical classification.
}
\details{
\tabular{ll}{
Package: \tab MKclass\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2019-09-20\cr
Depends: \tab R(>= 3.5.0)\cr
Imports: \tab stats\cr
Suggests: \tab knitr, rmarkdown, foreach, parallel, doParallel\cr
License: \tab LGPL-3\cr
URL: \tab http://www.stamats.de/\cr
}
library(MKclass)
}
\author{ Matthias Kohl \url{http://www.stamats.de}

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
